My second commit

+++ Backend/Controllers/reviewController.js
@@ -2,17 +2,22 @@ import siteSchema from "../Models/SiteData.js";
 import review from "../Models/review.js";
 import mongoose from "mongoose";
 
+
 export const saveReview = async(req,res)=>{
     try{
-        const ID = req.params.id;
-        const {rating, comment} = req.body;
+        const {ID, authorId} = req.params;
+        const {rating, comment} = req.body;  
 
-        const newReview = new review({ rating, comment });
+        if (!mongoose.Types.ObjectId.isValid(ID)) {
+            return res.status(500).json({ message: "Invalid site ID", type: "error" });
+        }
+
+        const newReview = new review({ rating, comment, author: authorId });
         await newReview.save();
         
         const site = await siteSchema.findById(ID);
         if (!site) {
-        return res.status(404).json({ message: "Site not found", type: "error" });
+        return res.status(404).json({ message: "Sorry, we couldn't find the palace you're looking for.", type: "error" });
         }
 
         site.review.push(newReview._id);
@@ -21,16 +26,20 @@ export const saveReview = async(req,res)=>{
         res.status(201).json({ message: "Thank you for your feedback", type: "success" });
     }catch(err){
         console.error("Error saving data:", err);
-        res.status(500).json({ message: "Server error", type: "error" });
+        res.status(500).json({ message: err.message || "Your review is not save(Server error)", type: "error" });
     }
 };
 
 export const fetchReview = async(req, res)=>{
     try{
         const ID = req.params.id;
+        if (!mongoose.Types.ObjectId.isValid(ID)) {
+            return res.status(500).json({ message: "Invalid site ID", type: "error" });
+        }
+
         const site = await siteSchema.findById(ID);
         if (!site) {
-            return res.status(404).json({ message: "Site not found", type: "error" });
+            return res.status(404).json({ message: "Sorry, we couldn't find the palace you're looking for.", type: "error" });
         }
         const reviewID = site.review;
         const reviewData = await review.find({ _id: { $in: reviewID } });
@@ -38,25 +47,36 @@ export const fetchReview = async(req, res)=>{
         res.status(200).json(reviewData);
     }catch(err){
         console.error("Error saving data:", err);
-        res.status(500).json({ message: err.message || "Server error", type: "error" });
+        res.status(500).json({ message: err.message || "Your review is not find(Server error)", type: "error" });
     }
 };
 
 export const deleteReview = async(req,res)=>{
     try{
         let reviewId = req.params.reviewId;
+        if (!mongoose.Types.ObjectId.isValid(reviewId)) {
+            return res.status(500).json({ message: "Invalid site ID", type: "error" });
+        }
         await review.findByIdAndDelete(reviewId);
+
+        await siteSchema.updateOne(
+            { review: reviewId },
+            { $pull: { review: reviewId } } // remove from array
+        );
 
         res.status(201).json({ message: "Your feedback delete successful", type: "success" });
     }catch(err){
         console.error("Error saving data:", err);
-        res.status(500).json({ message: err.message || "Server error", type: "error" });
+        res.status(500).json({ message: err.message || "Your review is not delete(Server error)", type: "error" });
     }
 };
 
 export const editReview = async(req, res)=>{
     try{
         let reviewId = req.params.reviewId;
+        if (!mongoose.Types.ObjectId.isValid(reviewId)) {
+            return res.status(500).json({ message: "Invalid site ID", type: "error" });
+        }
         const {rating, comment} = req.body;
 
         if (!rating || !comment) {
@@ -68,12 +88,27 @@ export const editReview = async(req, res)=>{
             { rating, comment },
         );
 
-        console.log(updatedReview);
         res.status(200).json({
             message: 'Review updated successfully.',
             type: "success"
         });
     }catch(err){
-        res.status(500).json({ message: err.message || "Server error" });
+        console.log(err);
+        res.status(500).json({ message: err.message ||"Your review is not update(Server error)", type: "error" });
+    }
+}
+
+export const findSingalReview = async(req,res)=>{
+    try{
+        const ID = req.params.reviewId;
+        if (!mongoose.Types.ObjectId.isValid(ID)) {
+            return res.status(500).json({ message: "Invalid site ID", type: "error" });
+        }
+        const reviewData = await review.findById(ID);
+        
+        res.status(200).json(reviewData);
+    }catch(err){
+        console.log(err);
+        res.status(500).json({ message: err.message || "Server error", type: "error" });
     }
 }
